src/main/java/com/dev/ajou/catchatgpt/service/MemberService.java
package com.dev.ajou.catchatgpt.service;

import com.dev.ajou.catchatgpt.domain.Member;
import com.dev.ajou.catchatgpt.repository.MemberRepository;

import java.util.List;
import java.util.Optional;

public class MemberService {

    private MemberRepository memberRepository;
    public MemberService(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    /*
    * 비즈니스 로직 작성
    */


    /*
    * 중복 회원 검사 -> 회원가입
    */
    public Long join(Member member) {
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    /*
     * 중복 회원 검사
     */
    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
                .ifPresent(m -> {
                    throw  new IllegalStateException("이미 존재하는 회원입니다.");
                });
    }

    /*
     * 전체 회원 조회
     */

    public List<Member> findMembers() {
        return memberRepository.findAll();
    }

    /*
     * 회원 한명 조회 (By ID)
     */
    public Optional<Member> findOne(Long memberId) {
        return memberRepository.findById(memberId);
    }

}
